exporter/requirements.txt
prometheus_client>=0.21.0
requests>=2.32.0
python-dotenv>=1.0.0
